+++ source/src/core/Movement.h
@@ -27,7 +27,7 @@ public:
 	void moveTo(Steps steps);
 	void stop();
 	void fullStop();
-	Steps rotateTo(float angle);
+	Steps rotateTo(double angle);
 
 private:
 	// ESP_FlexyStepper M1;
